a bit of logging to a beacon script

+++ beacon/openwrt.sh
@@ -18,17 +18,23 @@ configure_wifi() {
 }
 
 install_dependencies() {
-	opkg update
-	opkg install ncat
+	echo -e '\n### Checking dependencies =========================================\n' \
+	if [ ! -f /usr/bin/ncat ]
+	then
+		opkg update
+		opkg install ncat
+	fi
 }
 
 
 configure_ip() {
+	echo -e '\n### Configuring IP =========================================\n' \
 	ifconfig br-lan:0 192.168.4.2 up
 }
 
 configure_startup()
 {
+	echo -e '\n### Setting up startup script =========================================\n' \
 	local f=/etc/init.d/light-pong
 	echo '#!/bin/sh /etc/rc.common' > $f
 	echo 'START=99' >>$f
@@ -40,6 +46,7 @@ configure_startup()
 }
 
 start_broadcaster() {
+	echo -e '\n### Starting broker =========================================\n' \
 	ncat --broker -knl 2048
 }
